healthcare_chatbot_backend/requirements.txt
flask==2.0.3
flask-cors==3.0.10
PyPDF2==3.0.1
langchain==0.0.200
langchain-google-genai==0.0.6
langchain-community==0.0.38
numpy==1.26.4
pandas==1.5.3
torch==2.0.0
faiss-cpu==1.7.4
nltk==3.8.1
google-generativeai==0.3.2
python-dotenv==1.0.0
scikit-learn==1.0.2
joblib==1.1.0
cython==0.29.36
python-logstash==0.4.8
python-logstash-async==2.4.0
werkzeug==2.0.3

